requirements.txt
streamlit>=1.18.0
tensorflow>=2.12.0
pillow
numpy
streamlit-drawable-canvas
